# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventario Repuestos API
  version: 1.0.0
  description: API para autenticación, gestión de talleres, almacenes e invitaciones (code-based).
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      summary: Registro de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email o username ya registrado
  /api/auth/login:
    post:
      summary: Login con username o email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens y usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
  /api/auth/refresh:
    post:
      summary: Obtener nuevo access token usando refresh token
      operationId: apiAuthRefreshPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nuevo access token
  /api/auth/logout:
    post:
      summary: Revocar refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: OK
  /api/auth/oauth/google:
    post:
      summary: Login/registro con Google ID Token
      operationId: apiAuthOauthGooglePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthGoogleRequest'
      responses:
        '200':
          description: Tokens y usuario
  /api/auth/oauth/facebook:
    post:
      summary: Login/registro con Facebook access token
      operationId: apiAuthOauthFacebookPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthFacebookRequest'
      responses:
        '200':
          description: Tokens y usuario

  /api/auth/revoke-all:
    post:
      summary: Revocar todos los refresh tokens del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK

  /api/auth/me:
    get:
      summary: Obtener perfil del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized

  /api/talleres:
    post:
      summary: Crear taller (usuario autenticado es propietario)
      operationId: apiTalleresPost
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTallerRequest'
      responses:
        '201':
          description: Taller creado
    get:
      summary: Listar talleres del usuario autenticado
      operationId: apiTalleresGet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de talleres

  /api/talleres/{tallerId}/almacenes:
    post:
      summary: Crear almacen en taller
      operationId: apiTalleresTallerIdAlmacenesPost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tallerId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearAlmacenRequest'
      responses:
        '201':
          description: Almacen creado

  /api/talleres/{tallerId}/invitaciones/codigo:
    post:
      summary: Crear invitación por código para un taller
      operationId: apiTalleresTallerIdInvitacionesCodigoPost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tallerId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearInvitacionRequest'
      responses:
        '201':
          description: Código de invitación (raw code sólo se devuelve al creador)

  /api/talleres/invitaciones/accept:
    post:
      summary: Aceptar invitación por código
      operationId: apiTalleresInvitacionesAcceptPost
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequest'
      responses:
        '200':
          description: Usuario unido al taller

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - nombre
        - apellido
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        inviteCode:
          type: string
        # `rol` removed: role is assigned server-side and not accepted from clients
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        roles:
          type: array
          items:
            type: string
        activo:
          type: boolean
        fechaCreacion:
          type: string
          format: date-time
    LoginRequest:
      type: object
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
        device:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
    OAuthGoogleRequest:
      type: object
      properties:
        idToken:
          type: string
        inviteCode:
          type: string
        device:
          type: string
    OAuthFacebookRequest:
      type: object
      properties:
        accessToken:
          type: string
        inviteCode:
          type: string
        device:
          type: string
    CrearTallerRequest:
      type: object
      properties:
        nombre:
          type: string
    CrearAlmacenRequest:
      type: object
      properties:
        nombre:
          type: string
        ubicacion:
          type: string
    CrearInvitacionRequest:
      type: object
      properties:
        role:
          type: string
        daysValid:
          type: integer
    AcceptInvitationRequest:
      type: object
      properties:
        code:
          type: string